chore: major cypress upgrade from v9 to v12
