+++ Dockerfile.e2e
@@ -1,4 +1,4 @@
-FROM cypress/included:9.7.0
+FROM cypress/included:12.13.0
 
 # Copy and install node modules needed for frontend
 WORKDIR /e2e
