+++ apps/user-office-frontend-e2e/cypress/e2e/genericTemplates.cy.ts
