+++ apps/user-office-frontend-e2e/cypress/e2e/proposalAdministration.cy.ts
