+++ apps/user-office-frontend-e2e/cypress/support/e2e.ts
@@ -1,0 +1,27 @@
+// ***********************************************************
+// This example support/index.js is processed and
+// loaded automatically before your test files.
+//
+// This is a great place to put global configuration and
+// behavior that modifies Cypress.
+//
+// You can change the location of this file or turn off
+// automatically serving support files with the
+// 'supportFile' configuration option.
+//
+// You can read more here:
+// https://on.cypress.io/configuration
+// ***********************************************************
+
+// Import commands.js using ES2015 syntax:
+import './admin';
+import './call';
+import './instrument';
+import './proposal';
+import './template';
+import './user';
+import './utils';
+import './visits';
+import './workflow';
+import './sep';
+import './predefinedMessages';
