+++ apps/user-office-frontend-e2e/package.json
@@ -6,10 +6,10 @@
   "dependencies": {
     "@faker-js/faker": "^7.5.0",
     "@types/luxon": "^2.4.0",
-    "cypress": "^9.7.0",
-    "cypress-real-events": "^1.7.4",
+    "cypress": "^12.13.0",
+    "cypress-real-events": "^1.8.1",
     "graphql": "^16.6.0",
-    "graphql-request": "^5.0.0",
+    "graphql-request": "5.1.0",
     "graphql-tag": "^2.12.6",
     "jwt-decode": "^3.1.2",
     "luxon": "^2.5.0",
@@ -23,7 +23,7 @@
     "@typescript-eslint/parser": "^5.28.0",
     "eslint": "^8.17.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-cypress": "^2.12.1",
+    "eslint-plugin-cypress": "^2.13.3",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-unused-imports": "^2.0.0"
@@ -31,8 +31,8 @@
   "scripts": {
     "lint": "tsc --noEmit && eslint . --ext .js,.ts --quiet",
     "lint:fix": "tsc --noEmit && eslint . --ext .js,.ts --quiet --fix",
-    "cy:open": "cypress open --config defaultCommandTimeout=15000",
-    "cy:run": "cypress run --config defaultCommandTimeout=15000",
-    "cy:run:action": "cypress run --spec \"${CYPRES_SPEC_PATTERN}\" --browser chrome --config defaultCommandTimeout=15000"
+    "cy:open": "cypress open --config defaultCommandTimeout=15000 --e2e",
+    "cy:run": "cypress run --config defaultCommandTimeout=15000 --e2e",
+    "cy:run:action": "cypress run --spec \"${CYPRES_SPEC_PATTERN}\" --e2e --browser chrome --config defaultCommandTimeout=15000"
   }
 }
